Implemented volume control so that it works for Local & YouTube.

+++ src/me/oskarmendel/player/Player.java
@@ -70,9 +70,9 @@ public abstract class Player {
 	/**
 	 * Seeks the player to a new target time within the song or media. 
 	 * 
-	 * @param seekTime - Requested playback time.
+	 * @param seekTime - Requested playback time in seconds.
 	 */
-	public abstract void seek(double seekTime);
+	public abstract void seek(int seekTime);
 	
 	/**
 	 * Getter for the current playback time for the currently playing song
